[Feat3] Manejo de login de usuarios, guardar bitacora como json, validacion de admin en endpoints

+++ src/main/java/com/pruebaTecnica/api/Controller/ProductController.java
@@ -8,9 +8,14 @@ import com.pruebaTecnica.api.service.UserService;
 
 import lombok.AllArgsConstructor;
 
+import java.util.List;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+
+
+@CrossOrigin(origins = "http://localhost:4200")
 @RestController
 @AllArgsConstructor
 @RequestMapping("/products")
@@ -21,32 +26,72 @@ public class ProductController {
 
     private final String RoleErrorMsg =  "Este servicio solamente esta disponible para administradores";
 
-    @PostMapping
-    public ResponseEntity<Object> create(@RequestBody ProductDTO dto) {
-    
-        //boolean isAdmin = userService.getUserRole(idUser).equals("Admin");
-        //return isAdmin ? 
-         return  ResponseEntity.ok().body(productService.create(dto));
-        //    ResponseEntity.badRequest().body(RoleErrorMsg);    
+    @PostMapping("/user/{userId}")
+    public ResponseEntity<Object> create(@PathVariable Long userId,@RequestBody ProductDTO dto) {
+
+        boolean isAdmin = userService.getUserRole(userId).equals("Admin");
+
+        if (!isAdmin) {
+            return ResponseEntity.badRequest().body(RoleErrorMsg);
+        }
+
+        return ResponseEntity.ok().body(productService.create(dto));
     }
 
+    
     @PatchMapping("/{id}/user/{userId}")
-    public ProductEntity update(
-            @PathVariable Long id,
-            @PathVariable Long userId,
-            @RequestBody UpdateProductDTO dto) {
-        return productService.update(id, userId, dto);
+    public ResponseEntity<Object> update( @PathVariable Long id,
+                                          @PathVariable Long userId,
+                                          @RequestBody UpdateProductDTO dto ) {
+
+        boolean isAdmin = userService.getUserRole(userId).equals("Admin");
+
+        if (!isAdmin) {
+            return ResponseEntity.badRequest().body(RoleErrorMsg);
+        }
+
+        return ResponseEntity.ok(productService.update(id, userId, dto));
+    
     }
 
     @DeleteMapping("/{id}/user/{userId}")
-    public boolean delete(
-            @PathVariable Long id,
-            @PathVariable Long userId) {
-        return productService.delete(id, userId);
+    public ResponseEntity<Object> delete(@PathVariable Long id,
+                                         @PathVariable Long userId) {
+        
+        boolean isAdmin = userService.getUserRole(userId).equals("Admin");
+        
+        if (!isAdmin) {
+            return ResponseEntity.badRequest().body(RoleErrorMsg);
+        }
+        
+        boolean deleted = productService.delete(id, userId);
+        return ResponseEntity.ok(deleted);
     }
 
     @GetMapping("/{id}")
     public ProductEntity get(@PathVariable Long id) {
+
         return productService.get(id);
     }
+
+    @GetMapping("/getAll") 
+    public List<ProductEntity> getAll() {
+        return productService.getAll();
+    }
+
+    @GetMapping("/logs/user/{userId}")
+    public ResponseEntity<Object> getAllLogs(@PathVariable Long userId) {
+        
+        boolean isAdmin = userService.getUserRole(userId).equals("Admin");
+        
+        if (!isAdmin) {
+            return ResponseEntity.badRequest().body(RoleErrorMsg);
+        }
+    
+        return ResponseEntity.ok(productService.getAllLogs());
+
+    }
+    
+
+
 }
